Fix creation of user records after janitor trials
Previously, a negative "demoted from janitor trial" record would
always be created after a janitor trial ended, regardless of whether
or not the user in question was actually demoted or not.  This commit
removes the culprit, an overly aggressive after_destroy callback, and
instead directly invokes user record creation in the "demote!" method.